Add files via upload
Using the Raspberry Pi Pico W and additional components, I have written two scripts that need to be taken into account. "main" and "sequence", the alarm sequence will be imported into main so that the alarm system can run upon start up of the device or when the device receives power.