Add files via upload
With this commit, I am making changes towards the html script as it needed a slight touch up and to be further detailed. In addition, the button functionality has been corrected, both the on/off button. 

The next objective will be making these buttons work simultaneously whilst the other button is active, thinking to either break the while loop for it to return or to apply another set of if statements within the first if statement on line 94 - main.py.