Update README.md
Updating README with an issue log. I will constantly report any issue found within the project, and later update the README when I have addressed the problem and found a resolution.

Please note that any issues logged do not hinder the working capacity of the project as it is still able to function and perform from stages 1-3.